backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
